requirements/base.txt
amqp==5.3.1
anyio==4.12.0
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
billiard==4.2.4
black==25.11.0
celery==5.6.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
decorator==5.2.1
Django==5.2.9
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-filter==25.2
django-storages==1.14.6
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
exceptiongroup==1.3.1
executing==2.2.1
factory_boy==3.3.3
Faker==38.2.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
matplotlib-inline==0.2.1
mypy_extensions==1.1.0
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytokens==0.3.0
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.4
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
